Update for Onnx Convert() documentation so it is limited to current ONNX-ML target platforms compatibility
Afaik, current version of ML.NET Convert() to ONNX supports just ONNX-ML. This is an update for Onnx Convert() documentation so it is limited to ONNX-ML behavior for the mentioned examples. ONNX-ML is currently supported only by WinML. 
CoreML and additional Onnx Backends might come in the future, though.

ONNX-ML is part of the ONNX standard that provides functionality for classic ML and pipelines. It is a superset of core ONNX. Many frameworks like Caffe2 and CNTK will not support ONNX-ML since they are focused on DNN. 
AWS, Yahoo, and others have expressed interest in ONNX-ML but their products do not currently support it. This might change in the future. Export through ONNX-CoreML converter should be possible in the future, as well.